Remove richmenu API endpoints and consolidate rich menu scripts
