+++ README.md
@@ -173,7 +173,7 @@ npm run setup-richmenu:mock
 サーバー起動後に以下のエンドポイントにアクセスしてリッチメニューをセットアップすることもできます：
 
 ```
-GET /setup/richmenu
+# リッチメニューはスクリプトのみで完結（APIエンドポイントなし）
 ```
 
 注意: この方法を使用するには、有効なLINE APIの認証情報が必要です。
