+++ src/routes/api/index.ts
@@ -3,13 +3,7 @@ import { sendEveningNotification } from "../../features/notification/services/ev
 import { sendMorningNotification } from "../../features/notification/services/morning";
 import { asyncHandler } from "../../lib/errors";
 import { logger } from "../../lib/logger";
-import { setupRichMenu } from "../../services/richmenu";
 import type { ApiResponse } from "../../types";
-import {
-  generateThemedRichMenuImage,
-  PREDEFINED_THEMES,
-  saveRichMenuImageToTemp,
-} from "../../utils/richMenuImage";
 
 /**
  * ヘルスチェック用のエンドポイント
@@ -71,89 +65,6 @@ export const triggerEveningNotification = asyncHandler(
   },
 );
 
-/**
- * リッチメニューをセットアップするエンドポイント
- * @param c Honoコンテキスト
- * @returns レスポンス
- */
-export const setupRichMenuEndpoint = asyncHandler(
-  async (c: Context): Promise<Response> => {
-    logger.info("リッチメニューのセットアップを開始します...");
-
-    try {
-      // クエリパラメータからテーマを取得（デフォルトは "default"）
-      const theme = c.req.query("theme") || "default";
-      const saveTemp = c.req.query("save") === "true";
-
-      // テーマが有効かチェック
-      if (!Object.prototype.hasOwnProperty.call(PREDEFINED_THEMES, theme)) {
-        const response: ApiResponse = {
-          status: "error",
-          message: `無効なテーマです: ${theme}。利用可能なテーマ: ${Object.keys(PREDEFINED_THEMES).join(", ")}`,
-        };
-        return c.json(response, 400);
-      }
-
-      // テーマに基づいてリッチメニュー画像を生成
-      const imageBuffer = generateThemedRichMenuImage(theme);
-
-      // 一時ファイルとして保存（オプション）
-      let tempFilePath: string | undefined;
-      if (saveTemp) {
-        tempFilePath = saveRichMenuImageToTemp(
-          imageBuffer,
-          `richmenu-${theme}-${Date.now()}.png`,
-        );
-      }
-
-      // リッチメニューをセットアップ
-      const richMenuId = await setupRichMenu(imageBuffer);
-
-      const response: ApiResponse = {
-        status: "success",
-        message: `リッチメニューのセットアップが完了しました (テーマ: ${theme})`,
-        data: {
-          richMenuId,
-          theme,
-          tempFilePath,
-          availableThemes: Object.keys(PREDEFINED_THEMES),
-        },
-      };
-
-      return c.json(response);
-    } catch (error) {
-      logger.error("リッチメニューセットアップエラー", error);
-
-      const response: ApiResponse = {
-        status: "error",
-        message: `リッチメニューのセットアップに失敗しました: ${error instanceof Error ? error.message : "不明なエラー"}`,
-      };
-
-      return c.json(response, 500);
-    }
-  },
-);
-
-/**
- * 利用可能なリッチメニューテーマ一覧を取得するエンドポイント
- * @param c Honoコンテキスト
- * @returns レスポンス
- */
-export const getRichMenuThemes = asyncHandler(
-  async (c: Context): Promise<Response> => {
-    const response: ApiResponse = {
-      status: "success",
-      message: "利用可能なリッチメニューテーマ一覧",
-      data: {
-        themes: Object.keys(PREDEFINED_THEMES),
-        themeDetails: PREDEFINED_THEMES,
-      },
-    };
-
-    return c.json(response);
-  },
-);
-
 // APIルートをエクスポート
 export const apiRoutes = {
   health: {
@@ -165,14 +76,6 @@ export const apiRoutes = {
     },
     evening: {
       post: triggerEveningNotification,
-    },
-  },
-  richmenu: {
-    setup: {
-      get: setupRichMenuEndpoint,
-    },
-    themes: {
-      get: getRichMenuThemes,
     },
   },
 };
