+++ src/routes/index.ts
@@ -20,10 +20,6 @@ export const setupRoutes = (app: Hono): void => {
     app.post("/api/notification/morning", apiRoutes.notification.morning.post);
     app.post("/api/notification/evening", apiRoutes.notification.evening.post);
 
-    // リッチメニューAPIルート
-    app.get("/api/richmenu/setup", apiRoutes.richmenu.setup.get);
-    app.get("/api/richmenu/themes", apiRoutes.richmenu.themes.get);
-
     logger.info("ルートを設定しました");
   } catch (error) {
     logger.error("ルートの設定に失敗しました", error);
